mem0_server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mem0ai>=1.0.0
pydantic>=2.0.0

